alles outgesourced, warscheinlich geht irgendwas noch kaputt idk
