+++ bot.py
@@ -20,14 +20,12 @@ from dotenv import load_dotenv
 from tinydb import Query, TinyDB, where
 import tinydb
 from num2words import num2words
-
-from functions import ban, config, suggest, help, set, get, tools, leaderboard, stats
+from functions import ban, config, suggest, help, set, get, tools, leaderboard, stats, trio, mac, pppkkk
 
 load_dotenv()
 TOKEN = os.getenv("DISCORD_TOKEN")
 
 bot = commands.Bot(command_prefix=config.config['prefix'], help_command=None)
-
 
 @bot.command(name="leaderboard", description="show leaderboard")
 async def o(ctx):
@@ -50,721 +48,24 @@ async def b(ctx):
     await get.gget(ctx)
 
 @bot.command(name="vplan", description="get current vertretungsplan for your defined classes")
-async def vplan(ctx):
-    vplan.vplan(ctx)
+async def v(ctx):
+    await vplan.vplan(ctx)
 
-trio_running = False
 @bot.command(name="trio", description="play a game of trio")
 @commands.guild_only()
-async def trio(ctx, incount):
-    count = 0
-    try:
-        count = int(incount)
-    except Exception:
-        await ctx.send("%s ❌ wrong amount of rounds" % ctx.message.author.mention)
-        return
-    if count <= 0 or count > 20:
-        await ctx.send("%s ❌ noooo >:(" % ctx.message.author.mention)
-        return
+async def t(ctx, incount):
+    await trio.trio(ctx,incount)
 
-    global trio_running
-    if trio_running:
-        await ctx.send("%s some game is already running" % ctx.message.author.mention)
-        return
-
-    trio_running = True
-    await bot.change_presence(activity=discord.Game(name="Trio"))
-
-    category = bot.get_channel(802864419902062612)
-    await ctx.message.guild.create_text_channel("trio-tmp", category=category)
-    channel = discord.utils.get(ctx.guild.channels, name="trio-tmp")
-    ctx.channel = channel
-
-    time.sleep(3)
-
-    feld = [None] * 49
-
-    for x in range(0, 49):
-        feld[x] = random.randint(1, 9)
-
-    table = PrettyTable()
-
-    table.field_names = [" ", "A", "B", "C", "D", "E", "F", "G"]
-    side = "１２３４５６７"
-    for x in range(0, 49, 7):
-        table.add_row([side[x // 7]] + feld[x : x + 7])
-
-    table.hrules = ALL
-
-
-    user = None
-
-    punkte = []
-    ids = []
-
-
-    try:
-        os.remove("res/output.mp3")
-
-        musik = ["res/music1.mp3", "res/music2.mp3", "res/music3.mp3"]
-
-        random.shuffle(musik)
-
-        music = musik * 7
-
-        os.system(
-            'ffmpeg -i "concat:%s|%s|%s|%s|%s|%s|%s|%s|%s|%s" -acodec copy res/output.mp3'
-            % (
-                music[0],
-                music[1],
-                music[2],
-                music[3],
-                music[4],
-                music[5],
-                music[6],
-                music[7],
-                music[8],
-                music[9],
-            )
-        )
-
-    except:
-        print("ZAPPZARAPP ERROR MIT MUSIK KAPUTT")
-        pass
-
-    try:
-        voice_channel = ctx.author.voice.channel
-        vc = await voice_channel.connect()
-        source = discord.PCMVolumeTransformer(discord.FFmpegPCMAudio("res/output.mp3"))
-        vc.play(source)
-    except:
-        pass
-
-
-    suchs = []
-
-    while count != 0:
-
-        await bot.change_presence(activity=discord.Game(name="Trio"))
-        tab = "```python\n" + str(table) + "```"
-        such = random.randint(1, 50)
-        while such in suchs:
-            such = random.randint(1, 50)
-        suchs.append(such)
-        search = (
-            '🔎 SEARCHING FOR NUMBER: **%d**\n write "s" when you found a combination - you have **%s** seconds '
-            % (such, config.config['wait-timeout'])
-        )
-
-
-        embed = discord.Embed(color=0xF1A90F)
-        embed.add_field(name="Trio", value=tab, inline=False)
-        embed.add_field(name="> ", value=search, inline=False)
-        await ctx.send(embed=embed)
-
-        def check(m):
-            return m.channel == channel and m.content == "s"
-
-        def combocheck(m):
-            if not tools.valid(m):
-                print("NOT tools.valid")
-                return [False, None]
-
-            comb = m.split(" ")
-            c1 = comb[0]
-            c2 = comb[1]
-            c3 = comb[2]
-
-            if c1[0] == "g" or c1[0] == "G":
-                c1_x = 7
-            else:
-                c1_x = int(c1[0], 16) - 9
-
-            if c2[0] == "g" or c2[0] == "G":
-                c2_x = 7
-            else:
-                c2_x = int(c2[0], 16) - 9
-
-            if c3[0] == "g" or c3[0] == "G":
-                c3_x = 7
-            else:
-                c3_x = int(c3[0], 16) - 9
-
-            c1_index = (int(c1[1]) - 1) * 7 + int(c1_x) - 1
-            c2_index = (int(c2[1]) - 1) * 7 + int(c2_x) - 1
-            c3_index = (int(c3[1]) - 1) * 7 + int(c3_x) - 1
-            if tools.conti(c1_index, c2_index, c3_index):
-                if feld[c1_index] * feld[c2_index] + feld[c3_index] == such:
-                    antwort = "%s*%s+%s=%s" % (
-                        feld[c1_index],
-                        feld[c2_index],
-                        feld[c3_index],
-                        such,
-                    )
-                    return [True, antwort]
-                if feld[c1_index] * feld[c2_index] - feld[c3_index] == such:
-                    antwort = "%s*%s-%s=%s" % (
-                        feld[c1_index],
-                        feld[c2_index],
-                        feld[c3_index],
-                        such,
-                    )
-                    return [True, antwort]
-                if feld[c2_index] * feld[c3_index] + feld[c1_index] == such:
-                    antwort = "%s*%s+%s=%s" % (
-                        feld[c2_index],
-                        feld[c3_index],
-                        feld[c1_index],
-                        such,
-                    )
-                    return [True, antwort]
-                if feld[c2_index] * feld[c3_index] - feld[c1_index] == such:
-                    antwort = "%s*%s-%s=%s" % (
-                        feld[c2_index],
-                        feld[c3_index],
-                        feld[c1_index],
-                        such,
-                    )
-                    return [True, antwort]
-                if feld[c1_index] * feld[c3_index] + feld[c2_index] == such:
-                    antwort = "%s*%s+%s=%s" % (
-                        feld[c1_index],
-                        feld[c3_index],
-                        feld[c2_index],
-                        such,
-                    )
-                    return [True, antwort]
-                if feld[c1_index] * feld[c3_index] - feld[c2_index] == such:
-                    antwort = "%s*%s-%s=%s" % (
-                        feld[c1_index],
-                        feld[c3_index],
-                        feld[c2_index],
-                        such,
-                    )
-                    return [True, antwort]
-            return [False, None]
-
-        def trü(m):
-            return m.author.id == user and m.channel == channel
-
-        try:
-
-            if config.config['type-timeout'] != 15:
-                config.config['type-timeout'] = 15
-
-            msg = await bot.wait_for("message", check=check, timeout=config.config['wait-timeout'])
-
-            user = msg.author.id
-
-            multiple = ""
-
-            remove = False
-
-            try:
-                if punkte[ids.index(msg.author.id)] >= config.config['trio-double']:
-                    multiple = "\nyou need 2 combinations!"
-                    config.config['type-timeout'] += 10
-                    remove = True
-            except:
-                pass
-
-            try:
-                await ctx.send(
-                    "%s you have ⌚ **%s seconds** to send your solution... %s"
-                    % (msg.author.mention, config.config['type-timeout'], multiple)
-                )
-                msg = await bot.wait_for("message", check=trü, timeout=config.config['type-timeout'])
-            except asyncio.TimeoutError:
-                await ctx.send(
-                    "**%s 😔 no solution sent in time!**" % msg.author.mention
-                )
-                if msg.author.id in ids:
-                    punkte[ids.index(msg.author.id)] -= 1
-                else:
-                    ids += [msg.author.id]
-                    punkte += [-1]
-                count -= 1
-                continue
-
-            if remove:
-                config.config['type-timeout'] -= 10
-
-            feett = [True, None]
-            if multiple != "":
-                fetz = combocheck(str(msg.content)[0:8])
-                feett = combocheck(str(msg.content)[9:])
-            else:
-                fetz = combocheck(msg.content)
-
-            if fetz[0] and feett[0] and str(msg.content)[0:8] != str(msg.content)[9:]:
-                if multiple == "":
-                    await ctx.send(
-                        "%s %s is **correct!** ✅" % (msg.author.mention, fetz[1])
-                    )
-                else:
-                    await ctx.send(
-                        "%s %s is **correct!** ✅"
-                        % (
-                            msg.author.mention,
-                            (str(fetz[1]) + " & " + str(feett[1])).replace("*", "\*"),
-                        )
-                    )
-
-                if msg.author.id in ids:
-                    punkte[ids.index(msg.author.id)] += 1
-                    if punkte[ids.index(msg.author.id)] == config.config['trio-double']:
-                        await ctx.send(
-                            "%s you need 2 combinations from now on!"
-                            % (msg.author.mention)
-                        )
-                else:
-                    ids += [msg.author.id]
-                    punkte += [1]
-            else:
-                await ctx.send(
-                    "%s %s is **incorrect!** ❌" % (msg.author.mention, msg.content)
-                )
-                if msg.author.id in ids:
-                    punkte[ids.index(msg.author.id)] -= 1
-                else:
-                    ids += [msg.author.id]
-                    punkte += [-1]
-
-            count -= 1
-
-            time.sleep(2)
-
-        except asyncio.TimeoutError:
-            await ctx.send("**😔 no solution sent in time!**")
-            count -= 1
-
-
-    list = [x for _, x in sorted(zip(punkte, ids), key=lambda pair: pair[0])]
-    list.reverse()
-    place = 0
-    value = ""
-    if not list:
-        value = "."
-    for x in list:
-        medal = ""
-        if place == 0:
-            medal = ":first_place:"
-        if place == 1:
-            medal = ":second_place:"
-        if place == 2:
-            medal = ":third_place:"
-        value += "%s <@%s>: %s points\n" % (medal, x, punkte[ids.index(x)])
-        if len(list) > 1:
-            tools.to_lb(x, punkte[ids.index(x)], "trio")
-        place += 1
-
-    embed = discord.Embed(
-        title=":checkered_flag:  :ribbon:  FINISHED  :ribbon:  :checkered_flag:",
-        color=0xF1A90F,
-    )
-    embed.add_field(name="RANKING:", value=value, inline=False)
-    await ctx.send(embed=embed)
-
-
-    time.sleep(10)
-
-    await bot.change_presence(status=discord.Status.online)
-    await channel.delete()
-    try:
-        for x in bot.voice_clients:
-            await x.disconnect()
-    except:
-        pass
-
-    trio_running = False
-
-
-mac_running = False
 @bot.command(name="mac", description="play a nice game of mirroring and complementing")
 @commands.guild_only()
-async def mac(ctx, frange):
-
-    try:
-        frange = int(frange)
-        if not (frange == 100 or frange == 1000 or frange == 10000):
-            raise Exception
-    except:
-        await ctx.send(
-            "%s ❌ intools.valid range! (has to be 100,1000 or 10000)"
-            % ctx.message.author.mention
-        )
-        return
-
-    global mac_running
-    if mac_running:
-        await ctx.send(
-            "%s ❌ theres already a game running!" % ctx.message.author.mention
-        )
-        return
-
-    mac_running = True
-    await bot.change_presence(activity=discord.Game(name="mirroring and complementing"))
-
-    punkte = []
-    ids = []
-    category = bot.get_channel(802864419902062612)
-    await ctx.message.guild.create_text_channel("mac-tmp", category=category)
-    channel = discord.utils.get(ctx.guild.channels, name="mac-tmp")
-    ctx.channel = channel
-
-    time.sleep(3)
-    number = random.randint(1, frange - 1)
-
-    def mirror(kek, ln):
-        if ln == 2:
-            dec = int(kek / 10)
-            rst = kek - dec * 10
-            return rst * 10 + dec
-        elif ln == 3:
-            hun = int(kek / 100)
-            dec = int((kek - hun * 100) / 10)
-            rst = kek - hun * 100 - dec * 10
-            return rst * 100 + dec * 10 + hun
-        elif ln == 4:
-            tho = int(kek / 1000)
-            hun = int((kek - tho * 1000) / 100)
-            dec = int((kek - tho * 1000 - hun * 100) / 10)
-            rst = kek - tho * 1000 - hun * 100 - dec * 10
-            return rst * 1000 + dec * 100 + hun * 10 + tho
-        else:
-            return kek
-
-    if frange == 100:
-        flen = 2
-    elif frange == 1000:
-        flen = 3
-    elif frange == 10000:
-        flen = 4
-
-    chain = [number]
-
-    for i in range(1, 50):
-        if i % 2 == 0 and flen == 3:
-            number = 1000 - number
-        elif flen == 3 and i % 2 != 0:
-            number = mirror(number, 3)
-        elif i % 2 != 0 and flen == 2:
-            number = mirror(number, 2)
-        elif i % 2 == 0 and flen == 2:
-            number = 100 - number
-        elif i % 2 == 0 and flen == 4:
-            number = 10000 - number
-        elif i % 2 != 0 and flen == 4:
-            number = mirror(number, 4)
-        chain.append(number)
-
-    await ctx.send("**mirroring and complementing:**-")
-
-    progress = 1
-
-    def check(msg):
-        return msg.channel == ctx.channel
-
-    correct = True
-
-    while mac_running:
-        await bot.change_presence(
-            activity=discord.Game(name="mirroring and complementing")
-        )
-        if progress == 44 and frange != 10000:
-            number = random.randint(1, frange - 1)
-            await ctx.send("**starting with new number:...**")
-            progress = 1
-
-            chain = [number]
-
-            for i in range(1, 50):
-                if i % 2 == 0 and flen == 3:
-                    number = 1000 - number
-                elif flen == 3 and i % 2 != 0:
-                    number = mirror(number, 3)
-                elif i % 2 != 0 and flen == 2:
-                    number = mirror(number, 2)
-                elif i % 2 == 0 and flen == 2:
-                    number = 100 - number
-                elif i % 2 == 0 and flen == 4:
-                    number = 10000 - number
-                elif i % 2 != 0 and flen == 4:
-                    number = mirror(number, 4)
-                chain.append(number)
-
-        try:
-            time.sleep(0.5)
-            if correct:
-                if progress % 2 != 0:
-                    text = (
-                        ":white_square_button:   mirror: **%i**" % chain[progress - 1]
-                    )
-
-                    embed = discord.Embed(color=0xF1A90F, description=text)
-                    await ctx.send(embed=embed)
-
-                else:
-                    text = ":inbox_tray:   complement: **%i**" % chain[progress - 1]
-
-                    embed = discord.Embed(color=0xF1A90F, description=text)
-                    await ctx.send(embed=embed)
-
-
-            msg = await bot.wait_for("message", check=check, timeout=20)
-
-            try:
-                msg.content = str(int(msg.content))
-
-                if msg.content == str(chain[progress]):
-                    correct = True
-                    emoji = "\U00002705"
-                    await msg.add_reaction(emoji)
-                    if msg.author.id in ids:
-                        punkte[ids.index(msg.author.id)] = (
-                            int(punkte[ids.index(msg.author.id)]) + 1
-                        )
-                    else:
-                        ids += [msg.author.id]
-                        punkte += [1]
-                    progress += 1
-                else:
-                    emoji = "\U0000274C"
-                    await msg.add_reaction(emoji)
-                    correct = False
-            except:
-                correct = False
-                pass
-
-        except asyncio.TimeoutError:
-            break
-
-
-
-    list = [x for _, x in sorted(zip(punkte, ids), key=lambda pair: pair[0])]
-    list.reverse()
-    place = 0
-    value = ""
-    if not list:
-        value = "."
-    for x in list:
-        medal = ""
-        if place == 0:
-            medal = ":first_place:"
-        if place == 1:
-            medal = ":second_place:"
-        if place == 2:
-            medal = ":third_place:"
-        value += "%s <@%s>: %s points\n" % (medal, x, punkte[ids.index(x)])
-        if len(list) > 1:
-            tools.to_lb(x, punkte[ids.index(x)], "mac_%s" % frange)
-        place += 1
-
-    embed = discord.Embed(
-        title=":checkered_flag:  :ribbon:  FINISHED  :ribbon:  :checkered_flag:",
-        color=0xF1A90F,
-    )
-    embed.add_field(name="RANKING:", value=value, inline=False)
-    await ctx.send(embed=embed)
-
-    time.sleep(10)
-    await channel.delete()
-    await bot.change_presence(status=discord.Status.online)
-    mac_running = False
-
+async def m(ctx, frange):
+    await mac.mac(ctx.frange)
 
 pipapo = False
 @bot.command(name="pipapo", description="play pi-pa-po-ki-ka")
 @commands.guild_only()
 async def pppkkk(ctx):
-
-    ban = []
-
-    global pipapo
-    if pipapo:
-        await ctx.send(
-            "%s ❌ theres already a game running!" % ctx.message.author.mention
-        )
-        return
-
-    pipapo = True
-    await bot.change_presence(activity=discord.Game(name="pi-pa-po-ki-ka"))
-
-    punkte = []
-    ids = []
-    category = bot.get_channel(802864419902062612)
-    await ctx.message.guild.create_text_channel("pipapo-tmp", category=category)
-    channel = discord.utils.get(ctx.guild.channels, name="pipapo-tmp")
-    ctx.channel = channel
-
-
-    embed = discord.Embed(
-        title="pi-pa-po-ki-ka",
-        color=0xF1A90F,
-        description="pi: **2**\npa: **3**\n po: **5**\nki: **7**\nka: **11**",
-    )
-    await ctx.send(embed=embed)
-
-    chain = []
-    for i in range(1, 5000):
-        temp = ""
-        if i % 2 == 0:
-            temp += "pi"
-        if i % 3 == 0:
-            temp += "pa"
-        if i % 5 == 0:
-            temp += "po"
-        if i % 7 == 0:
-            temp += "ki"
-        if i % 11 == 0:
-            temp += "ka"
-        if temp == "":
-            temp = str(i)
-        chain.append(temp)
-
-    progress = 1
-
-    if progress == 4980:
-        await ctx.send("**Bruh, jetzt ist mal schluss**")
-        pipapo = False
-
-    def check(msg):
-        return msg.channel == ctx.channel and msg.author.id not in ban
-
-
-    await bot.change_presence(activity=discord.Game(name="pi-pa-po"))
-
-    while pipapo:
-
-        t = random.randint(1, 15)
-        time.sleep(t / 10)
-
-        captcha = random.randint(1, 50)
-
-        if captcha == 5 and progress != 1 and False:
-            r = str(random.randint(123, 123456))
-
-            list = [x for _, x in sorted(zip(punkte, ids), key=lambda pair: pair[0])]
-            list.reverse()
-            cap_id = list[0]
-
-            if cap_id in ban:
-                cap_id = list[1]
-
-            def cap(msg):
-                return (
-                    msg.channel == ctx.channel
-                    and msg.content == r
-                    and msg.author.id == cap_id
-                )
-
-            num = (
-                num2words(r)
-                .replace("q", "𝕢")
-                .replace("w", "𝕨")
-                .replace("e", "𝕖")
-                .replace("r", "𝕣")
-                .replace("t", "𝕥")
-                .replace("z", "𝕫")
-                .replace("u", "𝕦")
-                .replace("i", "𝕚")
-                .replace("o", "𝕠")
-            )
-            num = (
-                num.replace("p", "𝕡")
-                .replace("a", "𝕒")
-                .replace("m", "𝕞")
-                .replace("n", "𝕟")
-                .replace("b", "𝕓")
-                .replace("v", "𝕧")
-                .replace("c", "𝕔")
-                .replace("x", "𝕩")
-                .replace("y", "𝕪")
-            )
-            num = (
-                num.replace("l", "𝕝")
-                .replace("k", "𝕜")
-                .replace("j", "𝕛")
-                .replace("h", "𝕙")
-                .replace("g", "𝕘")
-                .replace("f", "𝕗")
-                .replace("d", "𝕕")
-                .replace("s", "𝕤")
-            )
-
-            await ctx.send(
-                "<@%s> please prove that you are alive by typing %s as a number, you have 15 seconds"
-                % (cap_id, num)
-            )
-            try:
-                await bot.wait_for("message", check=cap, timeout=15)
-                await ctx.send("not a bot... ok....")
-            except:
-                await ctx.send(
-                    "<@%s> you are now banned from this round!" % str(cap_id)
-                )
-                ban.append(cap_id)
-                punkte.remove(punkte[ids.index(cap_id)])
-                ids.remove(ids[ids.index(cap_id)])
-
-        try:
-
-            msg = await bot.wait_for("message", check=check, timeout=30)
-            while (msg.content).lower() != str(chain[progress - 1]):
-                msg = await bot.wait_for("message", check=check, timeout=30)
-
-            emoji = "\U00002705"
-            await msg.add_reaction(emoji)
-
-            if msg.author.id in ids:
-                punkte[ids.index(msg.author.id)] = (
-                    int(punkte[ids.index(msg.author.id)]) + 1
-                )
-            else:
-                ids += [msg.author.id]
-                punkte += [1]
-            progress += 1
-
-            if progress % 5 == 0:
-                embed = discord.Embed(
-                    title="Number: %s" % str(progress), color=0xF1A90F
-                )
-                await ctx.send(embed=embed)
-
-        except:
-            pipapo = False
-
-    list = [x for _, x in sorted(zip(punkte, ids), key=lambda pair: pair[0])]
-    list.reverse()
-    place = 0
-    value = ""
-    if not list:
-        value = "."
-    for x in list:
-        medal = ""
-        if place == 0:
-            medal = ":first_place:"
-        if place == 1:
-            medal = ":second_place:"
-        if place == 2:
-            medal = ":third_place:"
-        value += "%s <@%s>: %s points\n" % (medal, x, punkte[ids.index(x)])
-        if len(list) > 1 or len(ban) > 0:
-            tools.to_lb(x, punkte[ids.index(x)], "pipapo")
-        place += 1
-
-    embed = discord.Embed(
-        title=":checkered_flag:  :ribbon:  FINISHED  :ribbon:  :checkered_flag:",
-        color=0xF1A90F,
-    )
-    embed.add_field(name="RANKING:", value=value, inline=False)
-    await ctx.send(embed=embed)
-
-    time.sleep(10)
-    await channel.delete()
-    await bot.change_presence(status=discord.Status.online)
-    pipapo = False
-
+    await pppkkk.pppkkk(ctx)
 
 @bot.command(name="help")
 @commands.guild_only()
