+++ requirements.txt
@@ -8,3 +8,4 @@ prettytable==2.1.0
 protobuf==3.15.8
 python-dotenv==0.17.0
 tinydb==4.4.0
+ffmpeg==1.4.0
